feat: persisted Station ID passed to Zinnia (#424)
Introduce the concept of a Station instance ID that's represented as a public
key of an Ed25119 keypair.

Generate the keypair on the first start and store it in a state file. Load the
keypair from the state file on subsequent starts.

Introduce an optional env var PASSPHRASE allowing operators to specific a
passphrase that will be used to encrypt & decrypt the file storing the keypair.

Implement an upgrade path allowing users to start with an empty PASSPHRASE and
later upgrade their deployment with a PASSPHRASE string. Station will load the
plaintext keypair and encrypt it using the new PASSPHRASE when it detects this
situation.

Forward the Station ID to Zinnia via env vars and include it in the telemetry
data written to InfluxDB.

---------

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>
Co-authored-by: Patrick Nercessian <pjnercessian@gmail.com>